Removed hash functionality
There are so many libraries that can handle hash state.
Hash state should only be defined or given within a current
page. It should not be used to rebuild state and dispatch
*navigate* actions in order to get back to the state _after_
retrieving the state from the server resource.

It fights the purpose and does not correlate with this plugin.